Moved some of the block logic from Version into Version.block
