+++ PyMCTranslate/py3/api/version/translators/item.py
@@ -4,16 +4,21 @@ from PyMCTranslate.py3.api import Item, BlockItem
 from .base import BaseTranslator
 
 if TYPE_CHECKING:
-    from ..version import Version
+    from PyMCTranslate.py3.api.version import Version
+    from PyMCTranslate.py3.api.translation_manager import TranslationManager
 
 BlockCoordinates = Tuple[int, int, int]
 
 
 class ItemTranslator(BaseTranslator):
     def __init__(
-        self, parent_version: "Version", universal_format: "Version", database: dict
+            self,
+            translation_manager: "TranslationManager",
+            parent_version: "Version",
+            database: dict,
+            *_
     ):
-        super().__init__(parent_version, universal_format, database, "item")
+        super().__init__(translation_manager, parent_version, database, "item")
 
     def to_universal(
         self, object_input: Union[Item, BlockItem]
